Expose major tuning parameters.
Expose a bunch of major encoder tuning parameters through ctl's. Implement
needed parts in Opus and CELT. Add lowpass code to CELT. We will lower the
amount of coded bands if the lowpass is lower than a band boundary, and we
will zero out the MDCT coeffs above our cutoff. We reverse the procedure if
we detect that an extra frequency->time transformation is being done on our band.